requirements.txt
# AUPAT - Abandoned Upstate Project Archive Tool
# Python Dependencies
# Version: 1.0.0
# Last Updated: 2025-11-15

# Text Normalization
unidecode>=1.3.6         # Convert Unicode to ASCII for location names
python-dateutil>=2.8.2   # Date parsing and normalization
# postal>=1.1.10         # Address parsing (libpostal Python bindings) - OPTIONAL, requires system libpostal

# Testing Framework
pytest>=7.4.0            # Testing framework
pytest-cov>=4.1.0        # Coverage reporting
pytest-mock>=3.12.0      # Mocking for pytest
requests-mock>=1.11.0    # HTTP mocking for adapter tests

# Web Interface
Flask>=3.0.0             # Web framework for import interface

# API Integration (v0.1.2)
requests>=2.31.0         # HTTP client for Immich/ArchiveBox APIs
tenacity>=8.2.3          # Retry logic for API calls
Pillow>=10.0.0           # Image processing for dimensions extraction

# Note: Standard library modules used (no installation needed):
# - sqlite3 (database operations)
# - pathlib (path handling)
# - json (JSON parsing/generation)
# - hashlib (SHA256 hashing)
# - uuid (UUID generation)
# - logging (application logging)
# - argparse (CLI argument parsing)
# - datetime (timestamp handling)
# - shutil (file operations)

# External Tools Required (install separately):
# - exiftool: Image EXIF metadata extraction
#   macOS: brew install exiftool
#   Ubuntu/Debian: apt-get install libimage-exiftool-perl
#
# - ffmpeg/ffprobe: Video metadata extraction
#   macOS: brew install ffmpeg
#   Ubuntu/Debian: apt-get install ffmpeg
